pyproject.toml
[tool.poetry]
name = "contrack_api"
version = "0.1.0"
description = ""
authors = ["Rodrigo Castrejon <rcastrejon@dealergeek.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.2.0"
fastapi = "^0.79.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
requests = "^2.28.1"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
